src/nber_helper: fix 'count_digits_in_nber'.
In this method, a comparison was done between an int and a long in order
to count the digits of the number. So, for
numbers greater then INT_MAX, this comparison was never satisfied and we
were in an infinite loop.

+++ src/math/nber_helper.c
@@ -4,7 +4,7 @@
 
 unsigned int count_digits_in_nber(unsigned long n)
 {
-    unsigned int i = 1;
+    unsigned long i = 1;
     unsigned int compt = 0;
 
     while (n >= i) {
